Improve `adons::module`.
* Bring in more documentation for this addon.
* Move the extension to `World` into the addon rather than having
  it conditionally compiled within `core`.
* Let `flecs_stats` compile without `flecs_module`.
